handling code smell : repetitive code on result component
