+++ frontend/gaejosim/src/Component/CommonSearch/Result/Result.test.js
